Okay, big update
Updated the hover text on index.html so that it actually is in the right place (learned a bunch about transform)I wrote a bunch for the main blog page - a paragraph or so for each blog i think. I revised the index.html main text. Made a lot of stuff less wordy. Really focusing on that. yup