Added images for projects, added Modals for those images, and added alerts for when people click learn more
The alerts will notify users when they click to learn more about projects, that I will be added details and technologies used, soon. I do this by showing off alerts.

The modals were a bit compliated because it is a dynamic list that changes, along with it's picture. so the modal needs to consider this, and basically, needs to be in scope.